Rename `ImportCliftTypesPipe.h` file
The new name is `ImportCliftPipes.h`.
